feat(auth): auth-field (dry) & update formData to zod validated data
